worker/requirements.txt
opencv-python
pika
pillow
requests
websockets
kafka-python